Add edges one at a time when joining paths
It seems to be the case that when adding a wire to a MakeWire, rather
than adding all the edges simultaneously as one contigious wire, instead the code walks through the wire, and adds the edge segments one at a time.

This is bad, because the edges may be added in an order that makes the
resulting path discontinuous, which blows up.

Instead, walk through the edges using a WireExplorer, and add them in
order.

I still need to hammer out the edgecases, this commit is a WIP.
I'm also currently not sure I love the monoid instance on Paths, because
unconnectable paths can easilly introduce errors.
I think my initial plan was to automatically join non coincident paths
with a line segment, although even that is messy if the line segment
accidentally closes the Path